+++ CHANGELOG.md
@@ -9,6 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased] - ReleaseDate
 
+### bbtw
+
+#### New Features
+
+- `bbtw show` command for listing builds in a buildspace ([!185](https://gitlab.archlinux.org/archlinux/buildbtw/-/merge_requests/185))
+
 ## [0.0.5] - 2026-05-19
 
 ### backend